gaiacare/gaia_care/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
crispy-bootstrap5==2025.4
Django==5.0.1
django-allauth==65.8.0
django-crispy-forms==2.4
django-environ==0.12.0
django-filter==25.1
django-storages==1.14.6
django-widget-tweaks==1.5.0
dnspython==2.7.0
gunicorn==23.0.0
idna==3.10
Markdown==3.8
packaging==25.0
pillow==11.2.1
pip==25.0.1
pymongo==4.13.0
python-dateutil==2.9.0.post0
requests==2.32.3
resend==2.10.0
six==1.17.0
sqlparse==0.5.3
typing-extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
whitenoise==6.9.0
djongo==1.7.3
